fix parsing `-o -`
When parsing `-o -`, it parses as `o: true` because mri thinks that `-`
is the start of a new argument, _I think_. So to solve it, we pre
process the options to convert `-` to a fallback value, then parse and
then convert it back.

I wish we could use a symbol, but mri will crash.