Do not scan `tailwind-merge` sources for candidates (#16005)
Resolves #15722

This PR adds a list of ignored dependencies (in the current form only
`tailwind-merge`) to the Vite extension which, when included in the
dependency tree, are no longer scanned by the Tailwind CSS compiler for
candidates. This is to work around an issue where some dependencies
contain vast lists of valid Tailwind CSS class names which would
otherwise always be inlined in the build.

## Test plan

This was tested in our Vite playground on both dev and prod builds
across macOS and Windows:

### Windows prod build before

![telegram-cloud-photo-size-4-5823632138052944949-y](https://github.com/user-attachments/assets/a2bca1c7-cef3-4f04-a588-cbf1afdbdd37)

### Windows prod build after

This includes a debug `console.log(…)` to make sure it matches the right
module.

![telegram-cloud-photo-size-4-5823632138052944951-y](https://github.com/user-attachments/assets/109b82c5-1946-4d44-b812-d68474686eb7)

### Windows dev build after

This includes a debug `console.log(…)` to make sure it matches the right
module.

![telegram-cloud-photo-size-4-5823632138052944948-y](https://github.com/user-attachments/assets/41cbe7f1-f997-4c6c-b5fe-01d574f7a999)