implement different approach
Instead of trying to traverse ASTs and tracking unused variables in
various spots, let's reverse the algorithm.

ALWAYS emit everything, this has the added benefit that emitting CSS
variables and keyframes is only done once (in the `compile(…)` step, not
in each `build(…)` step).

However, the removal of the unused CSS variables is part of the
`optimizeAst` step. This has a few benefits:

1. All the logic is done in a single spot, instead of tracking all over
   the codebase.
2. We are already traversing for the optimization step, so we can hook
   in without additional walks of the entire AST.
3. It will be more correct, e.g.: if you have a valid utility, but
   invalid variant, since the utility is handled first it could be that
   we mark variables as used even though the no CSS will be generated
   due to the invalid variant.

   Since this the `optimizeAst` step never even sees the thrown-away
   variant+utility, it means that we don't even have to worry about
   this.

Had to add a context node, to different between `:root {}` coming from
`@theme` or coming from user CSS (which should stay untouched).