Standalone: Use bun baseline builds for Linux (#16244)
Closes #15181

This PR changes the Standalone builds to use bun baseline instead. This
compiles to a reduced instruction set and should work around the
compatibility issues experienced across older server hardware.

## Test plan

See
https://github.com/tailwindlabs/tailwindcss/issues/15181#issuecomment-2634621266

Also tested it with the repor from @npezza93 and it now works locally as
well with a Docker build.